feat(mesure): refactor actions
