+++ packages/api/src/routes/hasura-actions/mesure-etat/actions/index.js
@@ -1,0 +1,5 @@
+const actionMesureEtatUpsert = require("./actionMesureEtatUpsert");
+
+module.exports = {
+  actionMesureEtatUpsert,
+};
