+++ packages/api/src/routes/hasura-actions/mesure/actions/actionMesuresDelete.js
