+++ packages/api/src/routes/hasura-actions/mesure/actions/index.js
@@ -1,5 +1,5 @@
-const actionMesuresDelete = require("./hasura-actions.mesures.delete");
-const actionMesuresExport = require("./hasura-actions.mesures.export");
+const actionMesuresDelete = require("./actionMesuresDelete");
+const actionMesuresExport = require("./actionMesuresExport");
 
 module.exports = {
   actionMesuresDelete,
